+++ app/src/main/java/woowacourse/movie/provider/RepositoryProvider.kt
@@ -1,0 +1,20 @@
+package woowacourse.movie.provider
+
+import android.content.Context
+import woowacourse.movie.data.MovieTheaterDatabase
+import woowacourse.movie.repository.SettingRepository
+import woowacourse.movie.repository.TicketRepository
+
+object RepositoryProvider {
+    fun ticketRepository(context: Context): TicketRepository =
+        TicketRepository(
+            ticketDao(context),
+        )
+
+    fun settingRepository(context: Context): SettingRepository =
+        SettingRepository(
+            context.getSharedPreferences("setting", Context.MODE_PRIVATE),
+        )
+
+    fun ticketDao(context: Context) = MovieTheaterDatabase.db(context).ticketDao()
+}
