+++ app/src/main/java/woowacourse/movie/view/setting/SettingFragment.kt
@@ -4,17 +4,17 @@ import android.os.Bundle
 import android.view.View
 import android.widget.Toast
 import androidx.activity.result.contract.ActivityResultContracts.RequestPermission
-import woowacourse.movie.Provider
 import woowacourse.movie.R
 import woowacourse.movie.databinding.FragmentSettingBinding
 import woowacourse.movie.domain.model.Ticket
+import woowacourse.movie.provider.SettingProvider
 import woowacourse.movie.view.base.BaseFragment
 import woowacourse.movie.view.receiver.NotificationReceiver
 import java.time.LocalDateTime
 
 class SettingFragment : BaseFragment<FragmentSettingBinding>(R.layout.fragment_setting), SettingContract.View {
     private val presenter: SettingContract.Presenter by lazy {
-        Provider.settingPresenter(this)
+        SettingProvider.settingPresenter(this)
     }
     private val requestPermissionLauncher =
         registerForActivityResult(
